Split error mod between set and `Error` type; display `Error n/m: ` prefixes.
