+++ parser/src/error/etype.rs
@@ -1,0 +1,1 @@
+pub type Error = chumsky::error::Simple<char>;
