+++ parser/src/tests/corpus/negatives/bare-inquiry/expected
@@ -1,2 +1,1 @@
-
-found end of input
+Error 1/1: found end of input
