+++ parser/src/tests/corpus/negatives/invalid-let-syntax/expected
@@ -1,2 +1,1 @@
-
-found '\{' but one of .* was expected
+Error 1/1: found '\{' but one of .* was expected
